subprocess: move fork/exec implementation to separate file
Problem: The fork/exec specific process creation implementation is
intermingled with other "local" subprocess setup code, making the
local.c source file fairly large. This will only get worse when
other process creation implementations are introduced.

Move the fork/exec specific code to a new file, fork.c, to better
separate and abstract the process creation code.
